Combine volcano & kube batch tests
