+++ test/e2e/command.go
