+++ test/e2e/kube-batch/README.md
@@ -1,0 +1,4 @@
+These tests in this folder utilize the original ``Job`` object in kubernetes
+instead of our ``jobs.batch.volcano.sh`` object, once the depended features
+priorityClass is supported in our Job resource, they should be
+upgraded and enabled again.
